+++ packages/core/tests/main/rsbuild.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from '@rsbuild/core';
 import { pluginReact } from '@rsbuild/plugin-react';
-import { pluginWebExtend } from '../../../rsbuild-plugin/src/index.js';
 
 export default defineConfig({
-  plugins: [pluginReact(), pluginWebExtend()],
+  plugins: [pluginReact()],
 });
